add session stats DTO
